feat(devloop-cli-homologation): TASK-1968 — Promote the LLM backend catalog into the package + shim
